Record Component Detection: Per Species
It is a too strict assumption that all species must be identical
in their records. This commit adds a `species_record_components`
attribute created by `read_openPMD_params` that is a dictionary
for each species containing the usual list of record components.

I unified the naming a bit since it is useful if we stay with the
naming in the standard ("record component") and not introduce new
namings (such as "particle quantity").

Since the old attribute `avail_ptcl_quantities` is
- assuming all species are identical
- using a wrong naming

let us internally "deprecate" it for removal in the future.
Unfurtunately, I found no easy way to throw a warning on usage
since one would need to wrap the "__getattr[ibute]__" member for
that with a base class.

For the `interactive.py` class we should add an event handling
that is refreshing the `record components` of particles as soon
as a user selected a new `species`.

This can be done in an independent PR into `dev`.