Widget Events: Handle Senders
[ipywidget events](https://ipywidgets.readthedocs.org/en/latest/examples/Widget%20Events.html)
usually add either the *sender*, such as a button or dropdown widget,
or the *change* as an object (in sliders) to the function that is
called by its event trait.

To handle these, the parameter should be in the interface
(I called it `sender_change` to handle all unary event APIs) although
we do not depend on its actual value.

To reduce verbose code that only sets the `force` parameter of the
`refresh_field/ptcl` functions I introduce
[`functools.partial`](https://docs.python.org/2/library/functools.html#functools.partial)
that bind parameters for exactly such situations.