Python 3.4: Re-Enable
Python 3.4 was (accidentially) dropped in setup meta information
(which affects PyPI shipping and version selection). Also re-enables
testing for it in travis.

If there are no major problems with this release, we should support
Python 3.4 (although 3.5+ added major speed improvements) because
it's still the stable Python release on many distributions.