Fill Invalid Regions w/ Zero
Filling unwritten/invalid regions, e.g., access outside of written
ADIOS blocks, with not-a-number (NaN) created some confusion.
Technically, data in this region is literally not-a-number (yet), but
it seems easier to process data when it is just filled with zeros
instead.

We now always fill with zeros, hoping that "why is a region empty"
("because it was not written") does not create another kind of
confusion.

Side note: resizing during write of BTD is non-trivial (as in:
not supported by HDF5 and ADIOS), because we need to resize from
the "lower index end". HDF5 and ADIOS only support resizing the
upper index bound.